Reorganize index.js
